# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateRouteRequest:
    properties:
      name:
        type: string
      organization_id:
        type: string
    required:
    - name
    - organization_id
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: detailed error message
        type: string
    type: object
...
  dto.GetRouteDetailResponse:
    properties:
      active:
        type: boolean
      id:
        type: string
      name:
        type: string
      organization:
        $ref: '#/definitions/dto.OrganizationDetailResponse'
    type: object
...
  dto.ListItemRouteResponse:
    properties:
      active:
        type: boolean
      id:
        type: string
      name:
        type: string
      organization:
        $ref: '#/definitions/dto.OrganizationDetailResponse'
    type: object
...
  dto.ListRouteResponse:
    properties:
      current_page:
        type: integer
      routes:
        items:
          $ref: '#/definitions/dto.ListItemRouteResponse'
        type: array
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.OrganizationDetailResponse:
    properties:
      cnpj:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
      type:
        type: string
    type: object
...
  dto.SuccessMessageResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.UpdateRouteRequest:
    properties:
      name:
        type: string
      organization_id:
        type: string
    required:
    - name
    - organization_id
    type: object
...
host: localhost:8002
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Service for managing user identities, roles, and organizations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Uber dos Rios API
  version: v1.0.16
paths:
...
  /routes:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of routes
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListRouteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List routes
      tags:
      - routes
    post:
      consumes:
      - application/json
      description: Create a new route with the provided details
      parameters:
      - description: Route to create
        in: body
        name: route
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRouteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Route created successfully
          schema:
            $ref: '#/definitions/dto.SuccessMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new route
      tags:
      - routes
  /routes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific route by its ID
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Route deleted successfully
          schema:
            $ref: '#/definitions/dto.SuccessMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a route
      tags:
      - routes
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific route by its ID
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetRouteDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get route details
      tags:
      - routes
    put:
      consumes:
      - application/json
      description: Update the details of an existing route by its ID
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      - description: Route details to update
        in: body
        name: route
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRouteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Route updated successfully
          schema:
            $ref: '#/definitions/dto.SuccessMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update an existing route
      tags:
      - routes
  /routes/search:
    get:
      consumes:
      - application/json
      description: Search for routes based on a query string with pagination
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListRouteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Search routes
      tags:
      - routes
...
schemes:
- http
- https
swagger: "2.0"